+++ packages/gatsby-theme-aio/src/components/AnnouncementBlock/index.js
@@ -18,7 +18,8 @@ import PropTypes from 'prop-types';
 import { getElementChild, TABLET_SCREEN_WIDTH } from '../../utils';
 import classNames from 'classnames';
 
-const AnnouncementBlock = ({ className, heading, text, button, theme = 'light' }) => {
+const AnnouncementBlock = ({ className, heading, text, button, theme = 'light', variantsTypePrimary='primary', variantStyleOutline = "outline" }) => {
+
   const link = getElementChild(button);
 
   return (
@@ -65,7 +66,8 @@ const AnnouncementBlock = ({ className, heading, text, button, theme = 'light' }
             css={css`
               margin-top: var(--spectrum-global-dimension-size-200);
             `}>
-            <AnchorButton href={link.props.href} variant="accent" style="fill">
+            <AnchorButton href={link.props.href}  style={variantStyleOutline}
+              variant={variantsTypePrimary}>
               <span class="spectrum-Button-label">{link.props.children}</span>
             </AnchorButton>
           </div>