+++ packages/gatsby-theme-aio/src/components/Footer/index.js
@@ -18,7 +18,9 @@ import '@spectrum-css/typography';
 import { layoutColumns, getExternalLinkProps, DESKTOP_SCREEN_WIDTH, MOBILE_SCREEN_WIDTH } from '../../utils';
 import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
+import LinkedInRetargeting from './linkedInRetargetting'
 
+const DESKTOP_SCREEN_MAX_WIDTH = '1060px'; 
 const { APIs, services, community, support, developer, legal, allAPIs } = {
   allAPIs: {
     title: 'View all',
@@ -116,11 +118,12 @@ const { APIs, services, community, support, developer, legal, allAPIs } = {
       path: 'https://adobe.com/legal/terms.html'
     },
     {
-      title: 'Cookie preferences'
+      title: 'Cookie preferences',
+      path: '#/'
     },
     {
-      title: 'Do not sell my personal information',
-      path: 'https://adobe.com/privacy/ca-rights.html'
+      title: 'Do not sell or share my personal information',
+      path: 'https://adobe.com/privacy/us-rights.html'
     },
     {
       title: 'AdChoices',
@@ -130,6 +133,8 @@ const { APIs, services, community, support, developer, legal, allAPIs } = {
 };
 
 const OPEN_PRIVACY_ID = 'openPrivacy';
+
+const VIEW_ALL_APIS_DESC = 'View all APIs and Services';
 
 const Heading = ({ children }) => <h3 className="spectrum-Heading spectrum-Heading--sizeXS">{children}</h3>;
 
@@ -187,7 +192,7 @@ const Footer = ({ hasSideNav = false }) => (
       <div
         css={css`
           box-sizing: border-box;
-          max-width: ${layoutColumns(12)};
+          max-width: ${DESKTOP_SCREEN_MAX_WIDTH};
           margin: 0 auto;
           ${hasSideNav &&
           'margin: 0 var(--spectrum-global-dimension-size-800) 0 var(--spectrum-global-dimension-size-400)'};
@@ -224,7 +229,7 @@ const Footer = ({ hasSideNav = false }) => (
           css={css`
             display: grid;
             grid-template-areas: 'apis blogs support developer';
-            grid-template-columns: 30% 22% 19%;
+            grid-template-columns: 25%;
             gap: var(--spectrum-global-dimension-size-400);
 
             @media screen and (max-width: ${DESKTOP_SCREEN_WIDTH}) {
@@ -261,16 +266,26 @@ const Footer = ({ hasSideNav = false }) => (
                       </Link>
                     </li>
                   ))}
+                  {services.map(({ title, path }, i) => (
+                    <li key={i}>
+                      <Link isQuiet={true} variant="secondary">
+                        <a {...getExternalLinkProps(path)} href={path}>
+                          {title}
+                        </a>
+                      </Link>
+                    </li>
+                  ))}
                   <li>
                     <Link isQuiet={true}>
-                      <a {...getExternalLinkProps(allAPIs.path)} href={allAPIs.path}>
+                      <a {...getExternalLinkProps(allAPIs.path)} href={allAPIs.path} aria-labelledby="allAPIsDesc">
+                        <span id="allAPIsDesc" css={css`display: none;`}>{VIEW_ALL_APIS_DESC}</span>
                         <strong>{allAPIs.title}</strong>
                       </a>
                     </Link>
                   </li>
                 </List>
               </div>
-              <div
+              {/* <div
                 css={css`
                   margin-left: var(--spectrum-global-dimension-size-400);
                 `}>
@@ -285,7 +300,7 @@ const Footer = ({ hasSideNav = false }) => (
                     </li>
                   ))}
                 </ul>
-              </div>
+              </div> */}
             </div>
             <div
               css={css`
@@ -430,7 +445,7 @@ const Footer = ({ hasSideNav = false }) => (
                         />
                       </svg>
                       <Link isQuiet={true} variant="secondary">
-                        <a {...getExternalLinkProps(path)} href={path}>
+                        <a {...getExternalLinkProps(path)} href={path} aria-label={title}>
                           {title}
                         </a>
                       </Link>
@@ -442,7 +457,7 @@ const Footer = ({ hasSideNav = false }) => (
                   return (
                     <li key={i}>
                       <Link isQuiet={true} variant="secondary">
-                        <a id={OPEN_PRIVACY_ID} href="#" aria-hidden="true" focusable="false" tabindex="-1"></a>
+                        <a id={OPEN_PRIVACY_ID} href={path} aria-label="Cookie preferences" tabindex="0"></a>
                       </Link>
                     </li>
                   );
@@ -451,7 +466,7 @@ const Footer = ({ hasSideNav = false }) => (
                 return (
                   <li key={i}>
                     <Link isQuiet={true} variant="secondary">
-                      <a {...getExternalLinkProps(path)} href={path}>
+                      <a {...getExternalLinkProps(path)} href={path} aria-label={title}>
                         {title}
                       </a>
                     </Link>
@@ -476,6 +491,9 @@ const Footer = ({ hasSideNav = false }) => (
           </div>
         </div>
       </div>
+      {process.env.GATSBY_DC_LINKED_IN && (
+      <LinkedInRetargeting/>
+      )}
     </footer>
   </>
 );