+++ packages/gatsby-theme-aio/src/components/ProductCard/index.js
@@ -13,9 +13,10 @@
 import React, { cloneElement, useEffect, useState } from 'react';
 import { css } from '@emotion/react';
 import { HeroButtons } from '../Hero';
+import classNames from "classnames";
 import '@spectrum-css/typography';
 import '@spectrum-css/card';
-import { DESKTOP_SCREEN_WIDTH, TABLET_SCREEN_WIDTH } from '../../utils';
+import { DESKTOP_SCREEN_WIDTH, TABLET_SCREEN_WIDTH, MOBILE_SCREEN_WIDTH } from '../../utils';
 import PropTypes from 'prop-types';
 
 const counter = {
@@ -25,7 +26,7 @@ const counter = {
 };
 const alignMapping = ['flex-start', 'flex-end'];
 
-const ProductCard = ({ theme = 'lightest', width = '100%', icon, heading, text, buttons }) => {
+const ProductCard = ({ theme = 'lightest', width = '100%', icon, heading, text, buttons, className }) => {
   let initColumns = 100 / parseFloat(width);
 
   if (width === '33%') {
@@ -51,10 +52,10 @@ const ProductCard = ({ theme = 'lightest', width = '100%', icon, heading, text, 
   if (columns === 2 || columns === 3) {
     alignment = alignMapping[counter[columns] % columns] || 'center';
   }
-
+  const MIN_MOBILE_SCREEN_WIDTH = '320px'
   return (
     <section
-      className={`spectrum--${theme}`}
+      className={classNames(className, `spectrum--${theme}`)}
       css={css`
         display: inline-flex;
         flex-direction: column;
@@ -63,7 +64,7 @@ const ProductCard = ({ theme = 'lightest', width = '100%', icon, heading, text, 
         padding: var(--spectrum-global-dimension-size-400) 0;
         background: var(--spectrum-global-color-gray-100);
 
-        @media screen and (max-width: ${TABLET_SCREEN_WIDTH}) {
+        @media screen and (min-width: ${MIN_MOBILE_SCREEN_WIDTH}) and (max-width: ${MOBILE_SCREEN_WIDTH})  {
           display: flex;
           width: 100%;
           align-items: center;
@@ -71,13 +72,14 @@ const ProductCard = ({ theme = 'lightest', width = '100%', icon, heading, text, 
       `}>
       <div
         role="figure"
-        tabIndex="0"
+        // tabIndex="0"
         className="spectrum-Card spectrum-Card--sizeM"
         css={css`
           margin: 0 var(--spectrum-global-dimension-size-300);
           width: calc(var(--spectrum-global-dimension-size-4600) - var(--spectrum-global-dimension-size-800));
-          height: calc(var(--spectrum-global-dimension-size-4600) - var(--spectrum-global-dimension-size-500));
-
+          // height: calc(var(--spectrum-global-dimension-size-3600) - var(--spectrum-global-dimension-size-500));
+          height: var(--spectrum-global-dimension-size-3600);
+          
           &:hover {
             border-color: var(--spectrum-card-border-color, var(--spectrum-global-color-gray-200));
           }
@@ -153,6 +155,7 @@ const ProductCard = ({ theme = 'lightest', width = '100%', icon, heading, text, 
         <div className="spectrum-Card-footer">
           <HeroButtons
             buttons={buttons}
+            quiets={['outline', 'outline']}
             styles={['outline', 'outline']}
             variants={['secondary', 'accent']}
             css={css`
@@ -175,7 +178,8 @@ ProductCard.propTypes = {
   icon: PropTypes.element,
   heading: PropTypes.element,
   text: PropTypes.element,
-  buttons: PropTypes.element
+  buttons: PropTypes.element,
+  className: PropTypes.string
 };
 
 export { ProductCard };